+++ libs/@guardian/source-development-kitchen/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @guardian/source-development-kitchen
 
+## 7.0.1
+
+### Patch Changes
+
+- 80eea1d: Small refactors to allow Eslint updates
+
+## 7.0.0
+
+### Major Changes
+
+Updated peer dependency to `@guardian/source@7.0.0`
+
 ## 6.0.0
 
 ### Major Changes